Add plugin messaging
Plugin messaging are intended for packets which should not get their own packet type, they can also be used for messaging between the server and other client mods in the future.
Each plugin message has a name, this should be an unique name so it won't collide with other mods or plugins.

Example of server api

GameServer.SendPluginMessageToAll(name, data);

And listening for plugin messages

PacketEvents.OnPluginMessage.Add((c, message) =>
{
    if (message.Name == [...]
});